requirements.txt
# pyATS + Genie + Unicon
pyats[full]
# testing & helpers
pytest
jinja2
rich
PyYAML
